+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - PVP convenience functions `dtype_to_pvp_element` and `get_defined_pvp_dtype` in `sarkit.cphd`
+- PVP convenience functions `dtype_to_pvp_element` and `get_defined_pvp_dtype` in `sarkit.crsd`
+- PPP convenience functions `dtype_to_ppp_element` and `get_defined_ppp_dtype` in `sarkit.crsd`
 
 
 ## [1.9.0] - 2026-06-10